moved filterSecuredParams into the executor and made it public and static
