+++ src/com/qaprosoft/jenkins/pipeline/tools/maven/Maven.groovy
@@ -6,6 +6,8 @@ import java.util.regex.Matcher
 import com.qaprosoft.jenkins.pipeline.Configuration
 import com.qaprosoft.jenkins.Logger
 
+import static com.qaprosoft.jenkins.pipeline.Executor.*
+
 public class Maven {
     //TODO: migreate to traits as only it is supported in pipelines
     // https://issues.jenkins-ci.org/browse/JENKINS-46145
@@ -58,28 +60,6 @@ public class Maven {
             // Run the maven build
             buildGoals(goals)
         }
-    }
-    private def filterSecuredParams(goals) {
-        def arrayOfParmeters = goals.split()
-        def resultSpringOfParameters = ''
-        for (parameter in arrayOfParmeters) {
-            def resultString = ''
-            if (parameter.contains("token") || parameter.contains("TOKEN")) {
-                def arrayOfString = parameter.split("=")
-                resultString = arrayOfString[0] + "=********"
-            } else if (parameter.contains("-Dselenium_host")){
-                parameter = "-Dselenium_host=http://demo:demo@demo.qaprosoft.com:4444/wd/hub"
-                def pattern = "(\\-Dselenium_host=http:\\/\\/.+:)\\S+(@.+)"
-                Matcher matcher = Pattern.compile(pattern).matcher(parameter)
-                while (matcher.find()) {
-                    resultString = matcher.group(1) + "********" + matcher.group(2)
-                }
-            } else {
-                resultString = parameter
-            }
-            resultSpringOfParameters += resultString + ' '
-        }
-        return resultSpringOfParameters
     }
 
     public def buildGoals(goals) {
